+++ src/model/statistics.ts
@@ -7,12 +7,14 @@
 import { Dataset } from "./dataset";
 import { Variable } from "./variable";
 import { PlaceInfo } from "./place";
+import { CoordinateValues } from "@/states/controlState";
 
 export interface StatisticsSource {
   dataset: Dataset;
   variable: Variable;
   time: string | null;
   placeInfo: PlaceInfo;
+  coordinateValues: CoordinateValues;
 }
 
 export interface Histogram {